requirements.txt
asgiref==3.6.0
attrs==23.1.0
black==23.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
dj-database-url==1.3.0
Django==4.1.7
django-crispy-forms==2.0
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.1.1
psycopg2==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
whitenoise==6.4.0
